backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
supabase==2.3.1
openai==1.12.0
python-multipart==0.0.9
pydantic==2.6.1
langchain==0.1.9
langchain-openai==0.0.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2 